Add main class for spring application
