+++ matrasscraper/src/main/java/com/blackbook/matrasscraper/MatrasScraperApp.java
@@ -8,7 +8,6 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
  */
 @SpringBootApplication
 public class MatrasScraperApp {
-
     public static void main(String[] args) {
         SpringApplication.run(MatrasScraperApp.class, args);
     }
